Core,Tests: silence NU1605 warning
In order to use Visual Studio for building and debugging
FSharpLint, added `<NoWarn>$(NoWarn);NU1605</NoWarn>` to
Directory.Build.props file.

This has to be done because of version conflict as a result of
pinning of Microsoft.Build.* packages done in commit
57b4584 (Paket,Core,Tests: upgrade FCS to 41.0.1).

The errors due to version conflicts were like this:
```
Error NU1605 Warning As Error: Detected package downgrade: Microsoft.Build.Framework from 17.8.3 to 16.11.0. Reference the package directly from the project to select a different version.
 FSharpLint.Core -> Microsoft.Build.Tasks.Core 17.8.3 -> Microsoft.Build.Framework (>= 17.8.3)
 FSharpLint.Core -> Microsoft.Build.Framework (>= 16.11.0) FSharpLint.Core D:\Projects\work\FSharpLint\src\FSharpLint.Core\FSharpLint.Core.fsproj 1
```